+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 -   Added new metric "website is using green energy" [#14](https://github.com/Patr1ick/green-web-analyzer/issues/14)
+-   Added new metric "estimated website co2 emissions" [#15](https://github.com/Patr1ick/green-web-analyzer/issues/15)
 
 ## [0.2.5](https://github.com/Patr1ick/green-web-analyzer/releases/tag/0.2.5)